// apps/derrops-portal/src/client/derrops-cloud/models/update-api-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Derrops API
 * Internal API
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface UpdateApiDto {
    'name'?: string;
    'description'?: string;
    'externalUrl'?: string;
    'versionStrategy'?: UpdateApiDtoVersionStrategyEnum;
    'fetchCron'?: string;
    'fetchUrl'?: string;
}

export const UpdateApiDtoVersionStrategyEnum = {
    Manual: 'manual',
    UrlFetch: 'url_fetch'
} as const;

export type UpdateApiDtoVersionStrategyEnum = typeof UpdateApiDtoVersionStrategyEnum[keyof typeof UpdateApiDtoVersionStrategyEnum];